update each_vote_numers and unittest

+++ programs/jogo-vote/src/lib.rs
@@ -15,11 +15,11 @@ declare_id!("62TKbRM9rPsUu5kKeGUkoqP5M4CyLiVSogfaVT1uJFLQ");
 pub mod jogo_vote {
     use super::*;
 
-    pub fn initialize(ctx: Context<InitializeState>, max_vote_numbers: u64) -> Result<()> {
+    pub fn initialize(ctx: Context<InitializeState>, max_vote_numbers: u8) -> Result<()> {
         initialize_state(ctx, max_vote_numbers)
     }
 
-    pub fn vote(ctx: Context<Vote>, vote_number: u64) -> Result<()> {
+    pub fn vote(ctx: Context<Vote>, vote_number: u8) -> Result<()> {
         cast_vote(ctx, vote_number)
     }
 }
